net: move some Linux-specific tests to unixsock_linux_test.go file
Also changed name from TestUnix... to TestUnixgram....

Updates #21965

Reviewed-on: https://go-review.googlesource.com/66333
Run-TryBot: Ian Lance Taylor <iant@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
